rename: image -> item 이동 및 image 패키지 삭제
