+++ src/main/java/com/comehere/ssgserver/item/application/ItemService.java
@@ -6,8 +6,9 @@ import com.comehere.ssgserver.item.dto.req.ItemListReqDTO;
 import com.comehere.ssgserver.item.dto.req.ItemReqDTO;
 import com.comehere.ssgserver.item.dto.resp.ItemCalcRespDTO;
 import com.comehere.ssgserver.item.dto.resp.ItemDetailRespDTO;
+import com.comehere.ssgserver.item.dto.resp.ItemImageRespDTO;
 import com.comehere.ssgserver.item.dto.resp.ItemListRespDTO;
-import com.comehere.ssgserver.item.vo.req.ItemReqVO;
+import com.comehere.ssgserver.item.dto.resp.ItemThumbnailRespDTO;
 
 public interface ItemService {
 	ItemDetailRespDTO getItemDetail(Long id);
@@ -19,4 +20,8 @@ public interface ItemService {
 	ItemListRespDTO getItemList(ItemListReqDTO dto, Pageable page);
 
 	void createItem(ItemReqDTO dto);
+
+	ItemImageRespDTO getItemImages(Long itemId);
+
+	ItemThumbnailRespDTO getItemThumbnail(Long itemId);
 }
